+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   prior LSPS2 fee-limit state stored in `PaymentKind::Bolt11Jit` is not migrated.
 - Users of the VSS storage backend must upgrade their VSS server to at least version
   `v0.1.0-alpha.0` before upgrading LDK Node.
+- Usage of anchor channels now requires an Esplora or Electrum chain source that supports
+  `submitpackage`, or a Bitcoin Core RPC/REST chain source against Bitcoin Core v29 and above.
 
 # 0.7.0 - Dec. 3, 2025
 This seventh minor release introduces numerous new features, bug fixes, and API improvements. In particular, it adds support for channel Splicing, Async Payments, as well as sourcing chain data from a Bitcoin Core REST backend.